refactor(frontend): split R2 product route into modular files (1223→102 lines)
Extract loader, meta builder and main component from the monolithic route file
into dedicated modules for better maintainability and code navigation.

- pieces-vehicle.loader.server.ts: server-only loader (~340 lines)
- pieces-vehicle.meta.ts: meta tags + Schema.org JSON-LD (~110 lines)
- PiecesVehicleContent.tsx: render tree with hooks (~490 lines)
